Перемещение из модуля form всё что связано с фильтрами
