+++ source/js/form.js
@@ -7,6 +7,11 @@ import {
   renderAdvertisements
 } from './main.js';
 
+import {
+  deactivateFilters,
+  resetFilters
+} from './filter-advertisements.js'
+
 const TypePrices = {
   flat: 1000,
   bungalow: 0,
@@ -18,14 +23,6 @@ const MIN_NAME_LENGTH = 30;
 const MAX_NAME_LENGTH = 100;
 const MAX_PRICE = 1000000;
 
-const filtersMap = document.querySelector('.map__filters');
-const mapFeautures = filtersMap.querySelector('.map__features');
-const mapFilters = filtersMap.querySelectorAll('.map__filter');
-const housingType = filtersMap.querySelector('#housing-type');
-const housingPrice = filtersMap.querySelector('#housing-price');
-const housingRooms = filtersMap.querySelector('#housing-rooms');
-const housingGuests = filtersMap.querySelector('#housing-guests');
-const mapCheckboxes = document.querySelectorAll('.map__checkbox');
 const informForm = document.querySelector('.ad-form');
 const title = document.querySelector('#title');
 const address = document.querySelector('#address');
@@ -36,7 +33,7 @@ const timeout = document.querySelector('#timeout');
 const roomNumber = document.querySelector('#room_number');
 const capacity = document.querySelector('#capacity');
 const featureCheckboxes = document.querySelectorAll('.feature__checkbox')
-const informFieldsets = informForm.querySelectorAll('fieldset');
+const informFieldsets = document.querySelectorAll('fieldset');
 const description = document.querySelector('#description');
 const resetButton = document.querySelector('.ad-form__reset');
 const formPhoto = document.querySelector('.ad-form__photo')
@@ -47,12 +44,7 @@ const deactivateForm = () => {
   for (let i = 0; i < informFieldsets.length; i++) {
     informFieldsets[i].disabled = true;
   }
-
-  filtersMap.classList.add('ad-form--disabled');
-  mapFeautures.disabled = true;
-  for (let i = 0; i < mapFilters.length; i++) {
-    mapFilters[i].disabled = true;
-  }
+  deactivateFilters();
 }
 
 const activateForm = () => {
@@ -60,32 +52,6 @@ const activateForm = () => {
   for (let i = 0; i < informFieldsets.length; i++) {
     informFieldsets[i].disabled = false;
   }
-}
-
-const activateFilters = () => {
-  filtersMap.classList.remove('ad-form--disabled');
-  mapFeautures.disabled = false;
-  for (let i = 0; i < mapFilters.length; i++) {
-    mapFilters[i].disabled = false;
-  }
-}
-
-// Функция передает в коллбэк все values фильтров для карты
-const getFiltersValues= (cb) => {
-  const checkFeatures = () => {
-    let featuresValues = [];
-    for (let i = 0; i < mapCheckboxes.length; i++) {
-      if (mapCheckboxes[i].checked) {
-        featuresValues.push(mapCheckboxes[i].value)
-      }
-    }
-    return featuresValues;
-  }
-
-  filtersMap.addEventListener('change', ()=> {
-    cb(housingType.value,housingPrice.value,housingRooms.value,housingGuests.value,checkFeatures());
-  })
-
 }
 
 // validation
@@ -179,10 +145,6 @@ const onRoomNumberChange = function(){
 capacity.addEventListener('change', onRoomNumberChange);
 
 const resetForm = () => {
-  housingType.value = 'any';
-  housingPrice.value = 'any';
-  housingRooms.value = 'any';
-  housingGuests.value = 'any';
   title.value = '';
   address.value = '35.68170' + ', ' + '139.75388';
   type.value = 'flat';
@@ -197,7 +159,6 @@ const resetForm = () => {
 
   for (let i = 0; i < featureCheckboxes.length; i++) {
     featureCheckboxes[i].checked = false;
-    mapCheckboxes[i].checked = false;
   }
 
   mainPinMarker.setLatLng({
@@ -208,19 +169,23 @@ const resetForm = () => {
   renderAdvertisements();
 }
 
+const reset = () => {
+  resetForm();
+  resetFilters();
+}
+
 const onResetButtonClick = (evt) => {
   evt.preventDefault();
-  resetForm();
+  reset();
 }
 
 resetButton.addEventListener('click', onResetButtonClick);
-
 
 const setFormSubmit = () => {
   informForm.addEventListener('submit', (evt) => {
     evt.preventDefault();
     const formData = new FormData(informForm);
-    sendData(resetForm, formData)
+    sendData(reset, formData)
   });
 }
